Create README.md
Project Overview
This project focuses on applying clustering techniques to the Iris dataset using KMeans and Hierarchical Clustering algorithms. Clustering is a type of unsupervised learning where the goal is to group data points into clusters, such that data points within a cluster are more similar to each other than to those in other clusters. The Iris dataset consists of various features of iris flowers, and the task is to cluster the data into groups without using the target (species) information, which is useful for understanding natural groupings in the data.

Objective
The main objectives of this project are:

Loading and Preprocessing the Data:

Load the Iris dataset from sklearn.
Drop the species column as it's not used in clustering tasks.
Perform any necessary preprocessing steps.
Clustering Algorithm Implementation:

Implement KMeans Clustering.
Implement Hierarchical Clustering.
Visualize the clusters generated by each algorithm.
Model Evaluation and Comparison:

Analyze and compare the clustering results of both algorithms.
Dataset
The Iris dataset is a classic dataset in machine learning and contains measurements for three species of Iris flowers. The dataset consists of the following features:

Sepal length
Sepal width
Petal length
Petal width
The target variable, species, is not used in the clustering algorithms, as the goal is to identify the natural groupings of the data.

Setup
Prerequisites
To run this project, the following Python libraries need to be installed:

pandas
numpy
matplotlib
seaborn
scikit-learn